Resolved conflicts in:
- sui-indexer/src/handlers.rs
- sui-indexer/src/analyzer/oracle_manipulation.rs
- contracts/sources/defi_protocols/lending/compound_market.move
- contracts/tests/lending_tests.move

New features from main:
- create_shared_objects.sh for setting up shared objects
- setup.sh for contract deployment
- create_attack_tx.sh for generating attack transactions
- Enhanced detection logging with formatted output

+++ contracts/sources/utilities/math.move
@@ -72,9 +72,9 @@ public fun constant_product(reserve_a: u64, reserve_b: u64): u64 {
     reserve_a * reserve_b
 }
 
-/// Calculate square root using Babylonian method (simplified for u64)
+/// Calculate square root using Babylonian method (simplified for u128)
 /// For simulation purposes - not production-grade precision
-public fun sqrt(n: u64): u64 {
+public fun sqrt(n: u128): u64 {
     if (n == 0) {
         return 0
     };
@@ -88,5 +88,5 @@ public fun sqrt(n: u64): u64 {
         x = y;
         y = (x + n / x) / 2;
     };
-    x
+    (x as u64)
 }
